finish microbit-null demo and add some readme and code comments
