+++ demos/microbit/microbit-null/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "microbit-eusb"
+name = "microbit-null"
 version = "0.1.0"
 edition = "2021"
